Revert "Foreman environment orchestration"
This reverts commit a2575e0a3ec6fcaa4b74218b7db426e01354e4af.
