--- src/app/models/foreman/environment.rb
@@ -1,23 +1,0 @@
-#
-# Copyright 2012 Red Hat, Inc.
-#
-# This software is licensed to you under the GNU General Public
-# License as published by the Free Software Foundation; either version
-# 2 of the License (GPLv2) or (at your option) any later version.
-# There is NO WARRANTY for this software, express or implied,
-# including the implied warranties of MERCHANTABILITY,
-# NON-INFRINGEMENT, or FITNESS FOR A PARTICULAR PURPOSE. You should
-# have received a copy of GPLv2 along with this software; if not, see
-# http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
-
-module Foreman
-  class Environment < Resources::ForemanModel
-
-    attributes :name
-
-    def json_default_options
-      { :only => [:name] }
-    end
-
-  end
-end
