--- src/db/migrate/20121122110559_add_foreman_id_to_environment.rb
@@ -1,9 +1,0 @@
-class AddForemanIdToEnvironment < ActiveRecord::Migration
-  def self.up
-    add_column :environments, :foreman_id, :integer
-  end
-
-  def self.down
-    remove_column :environments, :foreman_id
-  end
-end
